Refactor the rest of BLANT functionalities into separate files
In addition to windowRep, code related to sampling, kovacs, output, synth graph, and utilities have been moved into their separate files. All source files have also been moved to the src folder instead of root. Makefile has been changed accordingly.
