+++ src/blant-synth-graph.c
@@ -1,0 +1,292 @@
+#include "blant-synth-graph.h"
+#include "blant-sampling.h"
+#include "blant-utils.h"
+#include "blant.h"
+#include "multisets.h"
+#include "heap.h"
+
+Boolean _GRAPH_GEN = false;
+int _GRAPH_GEN_EDGES;
+int _KS_NUMSAMPLES = 1000;
+float confidence; 
+int _genGraphMethod = -1;
+
+#if _GEN_SYN_GRAPH
+
+void reset_global_maps(int k_new) {_k = k_new; SetGlobalCanonMaps(); LoadMagicTable();}
+
+//Some Stats functions for later usage
+double* convertPDFtoCDF(double pdf[], double cdf[])
+{
+    int i;
+    cdf[0] = pdf[0];
+    for(i=1; i<_numConnectedCanon; i++) cdf[i] = cdf[i-1] + pdf[i] < 1 ? cdf[i-1] + pdf[i] : 1;
+    if((1-cdf[i-1]) < 0.0005) cdf[i-1] = 1;
+    if(cdf[i-1] != 1) 
+        printf("Increase rounding error margin. Last term of CDF is: %lf\n", cdf[i-1]);
+    assert(cdf[i-1] == 1);
+    return cdf;
+}
+
+double* copyConcentration(double src[], double dest[]) {for(int i=0; i<_numConnectedCanon; i++) dest[i] = src[i]; return dest;}
+
+double KStest(double empiricalCDF[], double theoreticalCDF[], int n)
+{
+    int i; 
+    double DMinus, DPlus, D, KS_stats;
+    D = -1;
+    for(i=0; i<_numConnectedCanon; i++) 
+    {
+        DPlus = empiricalCDF[i] - theoreticalCDF[i];
+        DMinus = theoreticalCDF[i] - empiricalCDF[i];
+        D = DPlus > D ? DPlus : D;
+        D = DMinus > D ? DMinus : D;
+    }
+    printf("Distance:  %lf\n", D);
+    KS_stats = (sqrt(n / 2.0) + 0.12 + 0.11 / sqrt(n / 2.0)) * D;
+    return KS_stats;
+}
+
+double KStestPVal(double KS_stats, float precision)
+{
+    double prev, curr, delta;
+    prev = 2 * pow(-1, 1-1) * exp(-2 * pow(1, 2) * pow(KS_stats, 2));
+    curr = prev + 2 * pow(-1, 2-1) * exp(-2 * pow(2, 2) * pow(KS_stats, 2));
+    delta = fabs(curr - prev);
+    int k = 3;
+    while (delta > precision)
+    {
+        prev = curr; curr = prev + 2 * pow(-1, k-1) * exp(-2 * pow(k, 2) * pow(KS_stats, 2));
+        delta = fabs(curr - prev);
+        k++;
+    }
+    return curr;
+}
+
+// Use subprocess (ForkBlant) to obtain concentration and index sampling results
+// Concentration is default to use MCMC sample Method and decimal frequency (concentration) output mode
+void LoadFromFork(int k, int numSamples, GRAPH* G, double onedarray[], double* twodarray[], int mode) 
+{
+    char line[_numOrbits * BUFSIZ], *pch, *tmp;
+    Boolean finished = false;
+    int i, j, canon, numRead, row, col, row_sum, total_sum;
+    double concentrationNum;
+    FILE * fpThread;
+    int Varray[_numConnectedCanon];
+
+    switch(mode)
+    {
+    case LOAD_CONCENTRATION:  // oned array is the connected cannonicals concentration [numConnected]
+        _outputMode = graphletFrequency; _freqDisplayMode = concentration; _sampleMethod = SAMPLE_MCMC;
+        fpThread = ForkBlant(k, numSamples, G);
+        for(i=0; i<_numConnectedCanon; i++) 
+        {
+            char *tmp = fgets(line, sizeof(line), fpThread); assert(tmp >= 0);
+            numRead = sscanf(line, "%lf %d", &concentrationNum, &canon);
+            assert(numRead == 2 && SetIn(_connectedCanonicals, canon));
+            onedarray[i] = concentrationNum;
+        }
+        break;
+    case LOAD_INDEX: // twodarray is of size [numSamples][k]
+        _outputMode = indexGraphlets; _sampleMethod = SAMPLE_NODE_EXPANSION;
+        fpThread = ForkBlant(k, numSamples, G);
+        for(i=0; i<numSamples; i++)
+        {
+            tmp = fgets(line, sizeof(line), fpThread); assert(tmp);
+            pch = strtok(line, " "); pch = strtok(NULL, " ");
+            for(j=0; j<k; j++) {twodarray[i][j] = (double)atoi(pch); pch = strtok(NULL, " ");}
+        }
+        break;
+    case LOAD_DISTRIBUTION:  
+    // onedarray is aggregated row concentration [numConnected]; 2darray is neighbor concentration [numConnected][numConnected]
+        _outputMode = graphletDistribution; _sampleMethod = SAMPLE_MCMC;
+        for(i=0; i<_numCanon; i++) 
+            memset(_graphletDistributionTable[i], 0, _numCanon);
+        fpThread = ForkBlant(k, numSamples, G);
+        row=0; col=0; row_sum=0; total_sum=0;
+        for(i=0; i<_numCanon; i++) {
+            tmp = fgets(line, sizeof(line), fpThread); assert(tmp);
+            if(SetIn(_connectedCanonicals, i)) {
+                pch = strtok(line, " ");
+                for(j=0; j<_numCanon; j++) {
+                    if(SetIn(_connectedCanonicals, j)) {twodarray[row][col++] += atoi(pch); row_sum += atoi(pch);}
+                    pch = strtok(NULL, " ");
+                }
+                total_sum += row_sum; onedarray[row++] = row_sum; row_sum=0; col=0; 
+            }
+        }
+       	assert(total_sum == numSamples);
+        for(i=0; i<_numConnectedCanon; i++) 
+        {
+            for(j=0; j<_numConnectedCanon; j++) 
+            	onedarray[i] != 0 ? twodarray[i][j] /= onedarray[i] : assert(twodarray[i][j] == 0);
+            onedarray[i] /= total_sum;
+        }
+        break;
+    default:
+        Fatal("Unknow loading mode");
+    }
+    if(feof(fpThread)) fclose(fpThread);
+}
+
+// Compare the synthetic graphlet with the original one by printing out each concentration results
+// Conduct a KS test and return the P-val (probability under NULL hypothethesis where the two distributions are the same)
+double compareSynGraph(GRAPH *G, GRAPH *G_Syn, int numSamples, int k, double theoreticalPDF[], double theoreticalCDF[])
+{
+    int i, canonArray[MAX_CANONICALS];
+    SetToArray(canonArray, _connectedCanonicals);
+    double KS_stats, P_val;
+    double empiricalPDF[_numConnectedCanon], empiricalCDF[_numConnectedCanon];
+    for(i=0;i<_numConnectedCanon; i++) {empiricalPDF[i]=0; empiricalCDF[i]=0;}
+    LoadFromFork(k, numSamples, G_Syn, empiricalPDF, NULL, LOAD_CONCENTRATION);
+    convertPDFtoCDF(empiricalPDF, empiricalCDF);
+    printf("GintOrdinal\tOriginalPDF\tSyntheticPDF\tOriginalCDF\tSyntheticCDF\n");
+    for(i=0; i<_numConnectedCanon; i++) printf("%i\t%lf\t%lf\t%lf\t%lf\n", canonArray[i], theoreticalPDF[i], empiricalPDF[i], theoreticalCDF[i], empiricalCDF[i]);
+    KS_stats = KStest(empiricalCDF, theoreticalCDF, numSamples);
+    P_val = KStestPVal(KS_stats, 0.0001);
+    printf("Obtained K_statistics:  %lf\n", KS_stats);
+    printf("Obtained p_value:   %lf\n\n", P_val);
+    return P_val;
+}
+
+// Generate a random Gint from concentration distribution
+// Return Gint in a binary adjacency matrix format
+int PickGraphletFromConcentration(int binaryNum[], double graphletCDF[], int k) 
+{
+    int canonArray[_numConnectedCanon];
+    SetToArray(canonArray, _connectedCanonicals);
+    int i, mid, l, h, GintOrdinal, Gint, step = 0;
+    do
+    {
+        double r = RandomUniform();
+        assert(0 <= r && r <= 1);
+        l=0, h=_numConnectedCanon-1;
+        while (l < h)
+        {
+            mid = l + ((h - l) >> 1);  // Same as mid = (l+h)/2
+            (r > graphletCDF[mid]) ? (l = mid + 1) : (h = mid);
+        }
+        GintOrdinal = (graphletCDF[l] >= r) ? l : -1;
+    } while(GintOrdinal < 0 && step < MAX_TRIES);
+    if(GintOrdinal < 0) Fatal("Unable to sample valid graphlet (GintOrdinal > 0) within MAX_TRIES");
+    GintOrdinal = canonArray[GintOrdinal];
+    Gint = _canonList[GintOrdinal];
+    int numBits = k * (k-1) / 2, n = Gint;
+    for(i=0; i<numBits; i++)
+    {
+        if(n>0) {binaryNum[i] = n % 2; n /= 2;} 
+        else binaryNum[i] = 0;
+    }
+    return Gint;
+}
+
+void stampFunction(GRAPH *G, int binaryNum[], int Varray[], int k)
+{
+    int i, j, z = 0, numBits = k * (k-1) / 2;
+    for(i=k-1; i>0; i--) for(j=i-1;j>=0;j--)
+    if(binaryNum[z++] == 1) 
+        GraphConnect(G, Varray[i], Varray[j]);
+    else 
+        GraphDisconnect(G, Varray[i], Varray[j]);
+   	assert(z == numBits);
+}
+
+// NBE-like synthetic generating method. 
+// Randomly pick out k node from G_SYN; Randomly pick out a Gint from G's concentration distribution
+// Stamp that Gint graphlet to the k nodes in G_SYN
+// Stop when number of edges of those two graphs are similar.
+// If KS test p-val ia smaller than the threshold and step < MAX_TRIES, delete half of the edges from G_SYN and repeat the steps above.
+void StampGraphletNBE(GRAPH *G, GRAPH *G_Syn, double graphletCDF[], int k, int k_small, double theoreticalPDF[], double theoreticalCDF[]) 
+{
+    double KS_stats, P_val;
+    int i, j, z, step, Varray[k], numBits = k*(k-1)/2, binaryNum[numBits], numRemoveSample, Gint, canonList_small[MAX_CANONICALS]; 
+    char BUF[BUFSIZ];
+    SET *connectedCanonicals_small = canonListPopulate(BUF, canonList_small, k_small);
+    SET *V = SetAlloc(G_Syn->n);
+    TINY_GRAPH *g = TinyGraphAlloc(k);
+    step=0, numRemoveSample=0;
+    do {
+        while(SetCardinality(V) < k) 
+            SetAdd(V, (int) G_Syn->n * RandomUniform());
+        assert(SetToArray(Varray, V) == k);
+
+        if(G_Syn->numEdges < _GRAPH_GEN_EDGES) {
+            Gint = PickGraphletFromConcentration(binaryNum, graphletCDF, k);
+            stampFunction(G_Syn, binaryNum, Varray, k);
+            TinyGraphInducedFromGraph(g, G_Syn, Varray);
+            assert(_K[Gint] == _K[TinyGraph2Int(g, k)]);
+        }
+        else {
+            printf("Steps made to remove Edges:  %i\n", numRemoveSample); 
+            step++; numRemoveSample=0;
+            reset_global_maps(k_small);
+            P_val = compareSynGraph(G, G_Syn, _KS_NUMSAMPLES, k_small, theoreticalPDF, theoreticalCDF);
+            reset_global_maps(k);
+            if(P_val > confidence || step >= MAX_TRIES) 
+                break;
+            while(G_Syn->numEdges > 0.5 * _GRAPH_GEN_EDGES) {
+                for(i=k-1; i>0; i--) for(j=i-1;j>=0;j--) 
+                    if(GraphAreConnected(G_Syn, Varray[i], Varray[j])) 
+                        GraphDisconnect(G_Syn, Varray[i], Varray[j]);
+
+                SetEmpty(V);
+                while(SetCardinality(V) < k) SetAdd(V, (int) G_Syn->n * RandomUniform());
+                assert(SetToArray(Varray, V) == k);
+                numRemoveSample++;
+            }
+        } 
+        SetEmpty(V);
+    } while(step < MAX_TRIES);
+    if(step > MAX_TRIES) printf("Steps Exceeding MAX_TRIES\n");
+    SetFree(V);
+}
+
+// Main function for generating synthetic graph
+int GenSynGraph(int k, int k_small, int numSamples, GRAPH *G, FILE *SynOutFile) 
+{
+    int i, j, Varray[MAX_CANONICALS];
+    double graphletPDF[_numConnectedCanon], graphletCDF[_numConnectedCanon], theoreticalPDF[_numConnectedCanon], theoreticalCDF[_numConnectedCanon], KS_stats;
+    double *distributionTablePDF[_numConnectedCanon], *distributionTableCDF[_numConnectedCanon];
+
+    GRAPH *G_Syn = GraphAlloc(G->n, SPARSE, _supportNodeNames);
+    switch(_genGraphMethod) 
+    {
+    case GEN_NODE_EXPANSION:
+        LoadFromFork(k, numSamples, G,  graphletPDF, NULL, LOAD_CONCENTRATION);
+        convertPDFtoCDF(graphletPDF, graphletCDF);
+        reset_global_maps(k_small);
+        LoadFromFork(k_small, numSamples, G,  theoreticalPDF, NULL, LOAD_CONCENTRATION);
+        convertPDFtoCDF(theoreticalPDF, theoreticalCDF);
+        reset_global_maps(k);
+        StampGraphletNBE(G, G_Syn, graphletCDF, k, k_small, theoreticalPDF, theoreticalCDF);
+        break;
+    case GEN_MCMC:
+        _graphletDistributionTable = Calloc(_numCanon, sizeof(int*));
+        for(i=0; i<_numCanon; i++) {
+            _graphletDistributionTable[i] = Calloc(_numCanon, sizeof(int));
+            for(j=0; j<_numCanon; j++) _graphletDistributionTable[i][j] = 0;
+        }
+        for(i=0; i<_numConnectedCanon; i++) {
+            distributionTablePDF[i] = Calloc(_numConnectedCanon, sizeof(double));
+            distributionTableCDF[i] = Calloc(_numConnectedCanon, sizeof(double));
+        }
+        LoadFromFork(k, numSamples, G, theoreticalPDF, distributionTablePDF, LOAD_DISTRIBUTION);
+        convertPDFtoCDF(theoreticalPDF, theoreticalCDF);
+        assert(theoreticalCDF[_numConnectedCanon-1] == 1.0);
+        for(i=0; i<_numConnectedCanon; i++) {
+            convertPDFtoCDF(distributionTablePDF[i], distributionTableCDF[i]);
+            assert((distributionTableCDF[i][_numConnectedCanon-1] - 1) < 0.00001 || distributionTableCDF[i][_numConnectedCanon-1] < 0.00001);
+        }
+        // StampGraphletMCMC(G, G_Syn, theoreticalCDF, distributionTableCDF, k);
+        break;
+    default:
+        Fatal("Unrecognized Synthetic Graph Generating method");
+    }
+    if(SynOutFile) {
+        GraphPrintConnections(SynOutFile, G_Syn);
+        fclose(SynOutFile);
+    }
+    return 0;
+}
+
+#endif
